+++ DataRepo/tests/management/commands/legacy/test_load_accucor_msruns.py
@@ -28,363 +28,8 @@ from DataRepo.models import (
 from DataRepo.models.study import Study
 from DataRepo.models.utilities import exists_in_db
 from DataRepo.tests.tracebase_test_case import TracebaseTestCase
-from DataRepo.utils import (
-    AggregatedErrors,
-    TracerLabeledElementNotFound,
-    read_from_file,
-)
-from DataRepo.utils.exceptions import (
-    DuplicatePeakGroup,
-    MultiplePeakGroupRepresentation,
-)
+from DataRepo.utils import read_from_file
 from DataRepo.utils.infusate_name_parser import parse_infusate_name_with_concs
-
-
-@override_settings(CACHES=settings.TEST_CACHES)
-class AccuCorDataLoadingTests(TracebaseTestCase):
-    fixtures = ["data_types.yaml", "data_formats.yaml", "lc_methods.yaml"]
-
-    @classmethod
-    def setUpTestData(cls):
-        call_command(
-            "load_studies",
-            infile=(
-                "DataRepo/data/tests/small_obob/"
-                "small_obob_animal_and_sample_table.xlsx"
-            ),
-        )
-        call_command(
-            "load_compounds",
-            infile=(
-                "DataRepo/data/tests/small_obob/"
-                "small_obob_animal_and_sample_table.xlsx"
-            ),
-        )
-        call_command(
-            "load_tracers",
-            infile=(
-                "DataRepo/data/tests/small_obob/"
-                "small_obob_animal_and_sample_table.xlsx"
-            ),
-        )
-        call_command(
-            "load_infusates",
-            infile=(
-                "DataRepo/data/tests/small_obob/"
-                "small_obob_animal_and_sample_table.xlsx"
-            ),
-        )
-        call_command(
-            "load_protocols",
-            infile=(
-                "DataRepo/data/tests/small_obob/"
-                "small_obob_animal_and_sample_table.xlsx"
-            ),
-        )
-        call_command(
-            "load_animals",
-            infile=(
-                "DataRepo/data/tests/small_obob/"
-                "small_obob_animal_and_sample_table.xlsx"
-            ),
-        )
-        call_command(
-            "load_tissues",
-            infile=(
-                "DataRepo/data/tests/small_obob/"
-                "small_obob_animal_and_sample_table.xlsx"
-            ),
-        )
-        call_command(
-            "load_samples",
-            infile=(
-                "DataRepo/data/tests/small_obob/"
-                "small_obob_animal_and_sample_table.xlsx"
-            ),
-        )
-        super().setUpTestData()
-
-    @classmethod
-    def load_glucose_data(cls):
-        """Load small_dataset Glucose data"""
-        call_command(
-            "legacy_load_accucor_msruns",
-            accucor_file="DataRepo/data/tests/small_obob/small_obob_maven_6eaas_inf_glucose.xlsx",
-            skip_samples=("blank"),
-            lc_protocol_name="polar-HILIC-25-min",
-            instrument="unknown",
-            date="2021-04-29",
-            researcher="Michael Neinast",
-            new_researcher=True,
-        )
-
-    def assure_coordinator_state_is_initialized(
-        self, msg="MaintainedModelCoordinators are in the default state."
-    ):
-        # Obtain all coordinators that exist
-        all_coordinators = [MaintainedModel._get_default_coordinator()]
-        all_coordinators.extend(MaintainedModel._get_coordinator_stack())
-        # Make sure there is only the default coordinator
-        self.assertEqual(
-            1, len(all_coordinators), msg=msg + "  The coordinator_stack is empty."
-        )
-        # Make sure that its mode is "always"
-        self.assertEqual(
-            "always",
-            all_coordinators[0].auto_update_mode,
-            msg=msg + "  Mode should be 'always'.",
-        )
-        # Make sure that the buffer is empty to start
-        for coordinator in all_coordinators:
-            self.assertEqual(
-                0, coordinator.buffer_size(), msg=msg + "  The buffer is empty."
-            )
-
-    @tag("multi-msrun")
-    def test_conflicting_peakgroups(self):
-        """
-        Test loading two conflicting PeakGroups rasies ConflictingValueErrors
-
-        Attempt to load two PeakGroups for the same Compound in the same MSRunSample
-        Note, when there are 2 different peak annotation files, that is an AmbiguousMSRuns error, but when other data
-        differs, it's a ConflictingValueErrors.  The formula for glucose was changed in the conflicting file.
-        """
-
-        self.load_glucose_data()
-
-        # The same PeakGroup, but from a different accucor file
-        with self.assertRaises(AggregatedErrors) as ar:
-            call_command(
-                "legacy_load_accucor_msruns",
-                accucor_file="DataRepo/data/tests/small_obob/small_obob_maven_6eaas_inf_glucose_conflicting.xlsx",
-                skip_samples=("blank"),
-                lc_protocol_name="polar-HILIC-25-min",
-                instrument="unknown",
-                date="2021-04-29",
-                researcher="Michael Neinast",
-                new_researcher=False,
-            )
-
-        aes = ar.exception
-        # This legacy loader doesn't summarize MultiplePeakGroupRepresentation exceptions
-        # 1 compounds, 2 samples -> 2 PeakGroups
-        self.assertEqual(2, aes.num_errors)
-        self.assertEqual(MultiplePeakGroupRepresentation, type(aes.exceptions[0]))
-
-    @tag("multi-msrun")
-    def test_duplicate_peak_group(self):
-        """Test inerting two identical PeakGroups raises an DuplicatePeakGroup error
-
-        This tests the AccuCorDataLoader.insert_peak_group method directly.
-        """
-
-        self.load_glucose_data()
-
-        # Setup an AccuCorDataLoader object with minimal info
-        # Required since using the "legacy_load_accucor_msruns" will not allow
-        # multiple loads of the same accucor_file, meaning two PeakGroups will
-        # differ in ArchiveFiles (peak annotation files) and raise ConflictingValueErrors, not DuplicatePeakGroup
-        adl = AccuCorDataLoader(
-            None,
-            None,
-            peak_annotation_file=None,
-            data_format=DataFormat.objects.get(code="accucor"),
-            date="2023-01-01",
-            researcher="",
-            lc_protocol_name="polar-HILIC-25-min",
-            instrument="",
-            peak_annotation_filename="peak_annotation_filename.tsv",
-            mzxml_files=[],
-        )
-        # Get the first PeakGroup, and collect attributes
-        peak_group = PeakGroup.objects.first()
-        peak_group_attrs = {
-            "name": peak_group.name,
-            "formula": peak_group.formula,
-            "compounds": peak_group.compounds,
-        }
-
-        # Test the instance method "insert_peak_group" rasies and error
-        # when inserting an exact duplicate PeakGroup
-        with self.assertRaises(Exception) as ar:
-            adl.insert_peak_group(
-                peak_group_attrs,
-                msrun_sample=peak_group.msrun_sample,
-                peak_annotation_file=peak_group.peak_annotation_file,
-            )
-        exc = ar.exception
-        self.assertEqual(
-            DuplicatePeakGroup,
-            type(exc),
-            msg=f"{DuplicatePeakGroup} expected. Got [{type(exc).__name__}: {exc}].",
-        )
-
-    @tag("multi-msrun")
-    def test_conflicting_peak_group(self):
-        """Test inserting two conflicting PeakGroups raises ConflictingValueErrors
-
-        Insert two PeakGroups that differ only in Formula.
-
-        This tests the AccuCorDataLoader.insert_peak_group method directly.
-        """
-
-        self.load_glucose_data()
-
-        # Setup an AccuCorDataLoader object with minimal info
-        adl = AccuCorDataLoader(
-            None,
-            None,
-            peak_annotation_file=None,
-            data_format=DataFormat.objects.get(code="accucor"),
-            date="2023-01-01",
-            researcher="",
-            lc_protocol_name="polar-HILIC-25-min",
-            instrument="",
-            peak_annotation_filename="peak_annotation_filename.tsv",
-            mzxml_files=[],
-        )
-        # Get the first PeakGroup, collect the attributes and change the formula
-        peak_group = PeakGroup.objects.first()
-        peak_group_attrs = {
-            "name": peak_group.name,
-            "formula": f"{peak_group.formula}S",
-            "compounds": peak_group.compounds,
-        }
-
-        with self.assertRaises(MultiplePeakGroupRepresentation):
-            adl.insert_peak_group(
-                peak_group_attrs,
-                msrun_sample=peak_group.msrun_sample,
-                peak_annotation_file=peak_group.peak_annotation_file,
-            )
-
-    def test_multiple_accucor_labels(self):
-        """
-        The infusate has tracers that cumulatively contain multiple Tracers/labels.  This tests that it loads without
-        error
-        """
-        call_command(
-            "legacy_load_animals_and_samples",
-            animal_and_sample_table_filename=(
-                "DataRepo/data/tests/accucor_with_multiple_labels/samples.xlsx"
-            ),
-        )
-        call_command(
-            "legacy_load_accucor_msruns",
-            accucor_file="DataRepo/data/tests/accucor_with_multiple_labels/accucor.xlsx",
-            lc_protocol_name="polar-HILIC-25-min",
-            instrument="unknown",
-            date="2021-04-29",
-            researcher="anonymous",
-            new_researcher=False,
-        )
-
-        # Test peak data labels
-        peak_data = PeakData.objects.filter(peak_group__name="Glycerol").filter(
-            peak_group__msrun_sample__sample__name="M1_mix1_T150"
-        )
-
-        peak_data_labels = []
-        for peakdata in peak_data.all():
-            pdl = peakdata.labels.values("element", "mass_number", "count")
-            peak_data_labels.append(list(pdl))
-
-        expected = [
-            [
-                {"element": "C", "mass_number": 13, "count": 0},
-            ],
-            [
-                {"element": "C", "mass_number": 13, "count": 1},
-            ],
-            [
-                {"element": "C", "mass_number": 13, "count": 2},
-            ],
-            [
-                {"element": "C", "mass_number": 13, "count": 3},
-            ],
-        ]
-
-        self.assertEqual(expected, list(peak_data_labels))
-
-    def test_accucor_bad_label(self):
-        """
-        This tests that a bad label in the accucor file (containing an element not in the tracers) generates a single
-        TracerLabeledElementNotFound error
-        """
-        call_command(
-            "legacy_load_animals_and_samples",
-            animal_and_sample_table_filename=(
-                "DataRepo/data/tests/accucor_with_multiple_labels/samples.xlsx"
-            ),
-        )
-        with self.assertRaises(AggregatedErrors) as ar:
-            call_command(
-                "legacy_load_accucor_msruns",
-                accucor_file="DataRepo/data/tests/accucor_with_multiple_labels/accucor_invalid_label.xlsx",
-                lc_protocol_name="polar-HILIC-25-min",
-                instrument="unknown",
-                date="2021-04-29",
-                researcher="anonymous",
-                new_researcher=False,
-            )
-        aes = ar.exception
-        self.assertEqual(1, len(aes.exceptions))
-        self.assertTrue(
-            isinstance(aes.exceptions[0], TracerLabeledElementNotFound),
-            msg="First exception must be TracerLabeledElementNotFound, but it was: "
-            f"[{type(aes.exceptions[0]).__name__}].",
-        )
-
-    @tag("multi-msrun")
-    def test_multiple_accucor_one_msrun(self):
-        """
-        Test that we can load different compounds in separate data files for the same sample run (MSRunSample)
-        """
-        self.load_glucose_data()
-        call_command(
-            "legacy_load_accucor_msruns",
-            accucor_file="DataRepo/data/tests/small_obob/small_obob_maven_6eaas_inf_lactate.xlsx",
-            lc_protocol_name="polar-HILIC-25-min",
-            instrument="unknown",
-            date="2021-04-29",
-            researcher="Michael Neinast",
-            new_researcher=False,
-        )
-        SAMPLES_COUNT = 2
-        PEAKDATA_ROWS = 11
-        MEASURED_COMPOUNDS_COUNT = 2  # Glucose and lactate
-
-        self.assertEqual(
-            PeakGroup.objects.count(), MEASURED_COMPOUNDS_COUNT * SAMPLES_COUNT
-        )
-        self.assertEqual(PeakData.objects.all().count(), PEAKDATA_ROWS * SAMPLES_COUNT)
-
-    @tag("multi-msrun")
-    def test_ambiguous_msruns_error(self):
-        """
-        Tests that an AmbiguousMSRuns exception is raised when a duplicate sample.peak group is encountered and the
-        peak annotation file names differ
-
-        This also tests that we do not allow the same compound to be measured from the
-        same sample run (MSRunSample) more than once
-        """
-        self.load_glucose_data()
-        with self.assertRaises(AggregatedErrors) as ar:
-            call_command(
-                "legacy_load_accucor_msruns",
-                # We just need a different file name with the same data, so _2 is a copy of the original
-                accucor_file="DataRepo/data/tests/small_obob/small_obob_maven_6eaas_inf_glucose_2.xlsx",
-                lc_protocol_name="polar-HILIC-25-min",
-                instrument="unknown",
-                date="2021-04-29",
-                researcher="Michael Neinast",
-                new_researcher=False,
-            )
-        # Check second file failed (duplicate compound)
-        aes = ar.exception
-        # This legacy loader doesn't summarize MultiplePeakGroupRepresentation exceptions
-        self.assertEqual(2, len(aes.exceptions))
-        self.assertTrue(isinstance(aes.exceptions[0], MultiplePeakGroupRepresentation))
 
 
 @override_settings(CACHES=settings.TEST_CACHES)
